Se agregaron las codificaciones para expresar las instrucciones en hexa, se agrego la libreria de entradas y salidas

+++ interrupciones.h
@@ -2,7 +2,7 @@
 #include "instrucciones.h"
 #include <stdint.h>
 
-void NVIC(uint32_t *registro,int *Banderas,uint8_t *SRAM,int *interrupciones,int *Indicador);
+void NVIC(uint32_t *registro,int *Banderas,uint8_t *SRAM,uint8_t *interrupciones,int *Indicador);
 
 void PUSH_LOCAL(uint32_t *registro,uint8_t *SRAM,uint8_t banderas);
